fix(migrations): no-issue: Moved test utilities to top level
